+++ src/main/java/act/handler/builtin/controller/impl/ReflectedHandlerInvoker.java
@@ -62,6 +62,7 @@ import org.osgl.inject.BeanSpec;
 import org.osgl.mvc.annotation.*;
 import org.osgl.mvc.result.*;
 import org.osgl.util.*;
+import org.w3c.dom.Document;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.*;
@@ -662,7 +663,7 @@ public class ReflectedHandlerInvoker<M extends HandlerMethodMetaInfo> extends Lo
     }
 
     private void ensureJsonDtoGenerated(ActionContext context) {
-        if (0 == fieldsAndParamsCount || !context.jsonEncoded()) {
+        if (0 == fieldsAndParamsCount || (!context.jsonEncoded() && !context.xmlEncoded())) {
             return;
         }
         Class<? extends JsonDto> dtoClass = jsonDTOClassManager.get(controllerClass, method);
@@ -710,6 +711,11 @@ public class ReflectedHandlerInvoker<M extends HandlerMethodMetaInfo> extends Lo
      */
     private String patchedJsonBody(ActionContext context) {
         String body = context.body();
+        if (context.xmlEncoded()) {
+            Document doc = XML.read(body);
+            JSONObject json = $.convert(doc).to(JSONObject.class);
+            body = JSON.toJSONString(json.containsKey("root") ? json.get("root") : json);
+        }
         if (S.blank(body) || 1 < fieldsAndParamsCount(context)) {
             return body;
         }