+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.8**
+* Basic support for xml content-type #905
 * ReflectedHandlerInvoker - try to get Annotation from method in parent class if possible #906
 * JsonView, CsvView etc shall be declared as `@Inherited` #904
 * Test debug page - make failed test display in the beginning of the page #896